Fix cache_only reason, post speech func split
There was two issues:
 - a missing final return in 'getTextInfoSpeech'
 - confusion about what the bool in the tuple returned
by 'getTextInfoSpeech' was for.

It seems the bool was used as "did speak", this can be gathered by
the contents of the sequence. So no longer return a tuple, just the speech
sequence. A key mistake was missing the fact that an empty sequence must be
returned from the end of the function. Now in this function, the 'retSequence'
variable is used in places where speech.speak was previously used. Only
'retSequence' should be returned, not the variable 'speechSequence'